// projects/svg-engine/edit/src/lib/nlu/types.ts
export type NluSlotSchema =
  | { readonly kind: 'number'; readonly optional?: boolean; readonly default?: number }
  | {
      readonly kind: 'color';
      readonly optional?: boolean;
      readonly default?: string;
    }
  | {
      readonly kind: 'enum';
      readonly values: readonly string[];
      readonly optional?: boolean;
      readonly default?: string;
    }
  | { readonly kind: 'string'; readonly optional?: boolean; readonly default?: string };